requirements.txt
hydra-core==1.3.2
imageio==2.31.1
matplotlib==3.7.1
networkx==3.3
numpy==1.23
omegaconf==2.3.0
overrides==7.3.1
pandas==1.4
pyemd==1.0.0
PyGSP==0.5.1
pytorch_lightning==2.0.4
scipy==1.13.1
setuptools==59.5.0
torch_geometric==2.3.1
torchmetrics==0.11.4
tqdm==4.65.0
wandb==0.15.4
rdkit==2023.9.1
seaborn==0.13.2
